+++ src/main/java/com/xti/jenkins/plugin/awslambda/upload/LambdaUploadVariables.java
@@ -69,6 +69,8 @@ public class LambdaUploadVariables extends AbstractDescribableImpl<LambdaUploadV
     private String subnets;
     private String securityGroups;
     private EnvironmentConfiguration environmentConfiguration;
+    private boolean enableDeadLetterQueue;
+    private String deadLetterQueueArn;
 
     @DataBoundConstructor
     public LambdaUploadVariables(String awsRegion, String functionName, String updateMode){
@@ -265,6 +267,24 @@ public class LambdaUploadVariables extends AbstractDescribableImpl<LambdaUploadV
         this.environmentConfiguration = environmentConfiguration;
     }
 
+    public boolean getEnableDeadLetterQueue() {
+        return enableDeadLetterQueue;
+    }
+
+    @DataBoundSetter
+    public void setEnableDeadLetterQueue(boolean enableDeadLetterQueue) {
+        this.enableDeadLetterQueue = enableDeadLetterQueue;
+    }
+
+    public String getDeadLetterQueueArn() {
+        return deadLetterQueueArn;
+    }
+
+    @DataBoundSetter
+    public void setDeadLetterQueueArn(String deadLetterQueueArn) {
+        this.deadLetterQueueArn = deadLetterQueueArn;
+    }
+
     public void expandVariables(EnvVars env) {
         awsAccessKeyId = ExpansionUtils.expand(awsAccessKeyId, env);
         clearTextAwsSecretKey = ExpansionUtils.expand(Secret.toString(Secret.fromString(awsSecretKey)), env);
@@ -283,6 +303,7 @@ public class LambdaUploadVariables extends AbstractDescribableImpl<LambdaUploadV
         if(environmentConfiguration != null){
             environmentConfiguration.expandVariables(env);
         }
+        deadLetterQueueArn = ExpansionUtils.expand(deadLetterQueueArn, env);
     }
 
     public LambdaUploadVariables getClone(){
@@ -306,6 +327,8 @@ public class LambdaUploadVariables extends AbstractDescribableImpl<LambdaUploadV
         if(environmentConfiguration != null){
             lambdaUploadVariables.setEnvironmentConfiguration(environmentConfiguration.getClone());
         }
+        lambdaUploadVariables.setEnableDeadLetterQueue(enableDeadLetterQueue);
+        lambdaUploadVariables.setDeadLetterQueueArn(deadLetterQueueArn);
         return lambdaUploadVariables;
     }
 
@@ -318,6 +341,9 @@ public class LambdaUploadVariables extends AbstractDescribableImpl<LambdaUploadV
             }
         } else {
             deployConfig.setEnvironmentVariables(new HashMap<String, String>());
+        }
+        if(enableDeadLetterQueue){
+            deployConfig.setDeadLetterQueueArn(deadLetterQueueArn);
         }
         return deployConfig;
     }