+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.5.8 (2017-01-11)
+
+Bugfixes:
+
+    - Adding empty lists in VPCConfig when not set to circumvent AWS backend change.
+
+## 0.5.7 (2017-01-02)
+
+Features:
+
+    - Added support for dead letter queue configuration
+
 ## 0.5.6 (2016-11-20)
 
 Features: